WIP changed nav-menu into container and presenter pattern
