+++ apps/coscrad-frontend/src/components/nav-menu/nav-menu-container.tsx
@@ -1,0 +1,53 @@
+import { CategorizableType } from '@coscrad/api-interfaces';
+import { useContext } from 'react';
+import { routes } from '../../app/routes/routes';
+import { ConfigurableContentContext } from '../../configurable-front-matter/configurable-content-provider';
+import { NavMenuPresenter } from './nav-menu-presenter';
+
+export type NavItemInfo = {
+    link: string;
+    label: string;
+};
+
+export const NavMenuContainer = (): JSX.Element => {
+    const { indexToDetailFlows } = useContext(ConfigurableContentContext);
+
+    // note this may be [] if we haven't included `notes`
+    const dynamicLinks = indexToDetailFlows
+        .filter(({ categorizableType }) => categorizableType === CategorizableType.note)
+        .map(({ label, route }) => ({
+            link: route || routes.notes.index,
+            label: label || 'Notes',
+        }));
+
+    // We may want an enum \ constants for our routes
+    const navItemInfos: NavItemInfo[] = [
+        {
+            link: routes.home,
+            label: 'Home',
+        },
+        {
+            link: routes.about,
+            label: 'About',
+        },
+        {
+            link: routes.resources.info,
+            label: 'Browse Resources',
+        },
+        {
+            link: routes.tags.index,
+            label: 'Tags',
+        },
+        {
+            link: routes.treeOfKnowledge,
+            label: 'Tree of Knowledge',
+        },
+        ...dynamicLinks,
+        {
+            link: routes.siteCredits,
+            label: 'Credits',
+        },
+    ];
+
+    return <NavMenuPresenter navItemInfos={navItemInfos} />;
+};
